pyproject.toml
[tool.poetry]
name = "vacancy-parsing"
version = "0.1.0"
description = ""
authors = ["kirill_tarasov <kirilltaras1792@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.4"


[tool.poetry.group.lint.dependencies]
mypy = "^1.8.0"
black = "^23.12.1"
isort = "^5.13.2"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
